requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
redis==5.0.0
Pillow==10.0.0